Added a system for media handlers
The gallery can now have multiple handlers installed, one handler is
chosen per slide depending on what they support. It is up to the handler
the initialize the slide markup and provide a thumbnail url.

This allows the decoupling of the video code to the lg-video plugin.
In addition this will allow external code to extend the gallery by
providing their own handlers for other video formats.

Promises are used in communication with the handlers to ensure that
async code can still work.

Each handler is registered in $.fn.lightGallery.mediaHandlers and
activated in $.fn.lightGallery.mediaHandlersActive.
The handler must have a "static" method called "match" which is called
for each slide, if the src url is known the method must instantiate
the handler and return it.

The following methods are available:
- init - Called directly after handler has been created and stored on
         slide
- thumbnail - Figures out thumbnail for slide and sends the result in
              the promise object.
- loadContent - Called for setting up the initial markup of the slide.
                Once finished the promise must be resolved.